+++ src/containers/Login/FingerprintPopup/ShakingText.component.js
