+++ src/containers/Login/FingerprintPopup/assets/finger_print@2x.png
