+++ src/containers/Login/FingerprintPopup/assets/finger_print@3x.png
